% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{bar}
\alias{bar}
\alias{bar,ANY-method}
\title{Bar}
\usage{
bar(x)

\S4method{bar}{ANY}(x)
}
\arguments{
\item{x}{\strong{Signature argument}.
Main input argument.}

\item{x}{\code{\link{ANY}}.}
}
\value{
\code{\link{character}}. \code{"Hello World!"}
}
\description{
Just for illustration purposes, this is a generic for which \emph{this}
package also provides a method.
}

