% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{bar}
\alias{bar}
\alias{bar,ANY-method}
\title{Bar (Generic With Default Method)}
\usage{
bar(x, ...)

\S4method{bar}{ANY}(x, ...)
}
\arguments{
\item{x}{\strong{Signature argument}.
Main input argument.}

\item{...}{Further arguments to be passed to subsequent functions.}

\item{x}{\code{\link{ANY}}.}
}
\value{
\code{\link{character}}. \code{"Hello World!"}
}
\description{
Just for illustration purposes, this is a generic for which \emph{this}
package also provides a method.
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/roxygen2-problem}
}

