% Generated by roxygen2 (4.0.0): do not edit by hand
\name{foo1}
\alias{foo1}
\title{Foo 1}
\usage{
foo1(x)
}
\arguments{
\item{x}{\strong{Signature argument}. Main input argument.}
}
\description{
I would like to export this as a mere interface for \emph{other} packages.
There shall be no default method that will be provided by \emph{this} package.
}

