% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bar}
\alias{bar}
\title{Bar}
\usage{
bar(x)
}
\arguments{
\item{x}{\strong{Signature argument}.
Main input argument.}
}
\description{
Just for illustration purposes, this is a generic for which \emph{this}
package also provides a method.
}

