% Generated by roxygen2 (4.0.0): do not edit by hand
\name{foo}
\alias{foo}
\title{Foo (Generic Only)}
\usage{
foo(x, ...)
}
\arguments{
\item{x}{\strong{Signature argument}.
Main input argument.}

\item{...}{Further arguments to be passed to subsequent functions.}
}
\description{
I would like to export this as a mere interface for \emph{other} packages.
There shall be no default method that will be provided by \emph{this} package.
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/roxygen2-problem}
}

